Move AudioManager to its own class file
